graph_utils: move Colormap to a dedicated module and rename it to ColormapFactory
