+++ bcdi/postprocessing/postprocessing_runner.py
@@ -23,6 +23,7 @@ import pprint
 import tkinter as tk
 from tkinter import filedialog
 
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 from bcdi.experiment.detector import create_roi
 from bcdi.experiment.setup import Setup
@@ -135,8 +136,7 @@ def run(prm):
         bad_color = "0.7"
     else:
         bad_color = "1.0"  # white background
-    colormap = gu.Colormap(bad_color=bad_color)
-    my_cmap = colormap.cmap
+    my_cmap = ColormapFactory(bad_color=bad_color).generate_cmap()
 
     #################################
     # define the experimental setup #
