+++ bcdi/utils/validation.py
@@ -11,6 +11,22 @@ from numbers import Real, Number
 import numpy as np
 
 
+def is_float(string):
+    """
+    Return True is the string represents a number.
+
+    :param string: the string to be checked
+    :return: True of False
+    """
+    if not isinstance(string, str):
+        raise TypeError("the input should be a string")
+    try:
+        float(string)
+        return True
+    except ValueError:
+        return False
+
+
 def valid_container(
     obj,
     container_types,
