+++ scripts/graph/bcdi_visu_2D_slice.py
@@ -12,6 +12,8 @@ import matplotlib
 import matplotlib.pyplot as plt
 import tkinter as tk
 from tkinter import filedialog
+
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 import bcdi.utils.utilities as util
 
@@ -72,8 +74,7 @@ if grey_background:
     bad_color = "0.7"
 else:
     bad_color = "1.0"  # white background
-colormap = gu.Colormap(bad_color=bad_color)
-my_cmap = colormap.cmap
+my_cmap = ColormapFactory(bad_color=bad_color).generate_cmap()
 
 if field in {"angle", "modulus"}:
     scale = "linear"
