+++ scripts/postprocessing/bcdi_prtf_BCDI_2D.py
@@ -21,6 +21,8 @@ from tkinter import filedialog
 import xrayutilities as xu
 import gc
 import sys
+
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 from bcdi.experiment.setup import Setup
 import bcdi.utils.image_registration as reg
@@ -202,8 +204,7 @@ hxrd = xu.experiment.HXRD(sample_inplane, sample_outofplane, qconv=qconv)
 ###################
 # define colormap #
 ###################
-colormap = gu.Colormap()
-my_cmap = colormap.cmap
+my_cmap = ColormapFactory().generate_cmap()
 
 ##########################################################
 # load experimental data, extracted 2D slice and 2D mask #
