+++ scripts/postprocessing/bcdi_strain_mean_var_rms.py
@@ -11,7 +11,9 @@ import numpy as np
 from matplotlib import pyplot as plt
 import tkinter as tk
 from tkinter import filedialog
+
 import bcdi.postprocessing.postprocessing_utils as pu
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 
 helptext = """
@@ -36,8 +38,7 @@ debug = True  # True to see data plots
 # define colormap #
 ###################
 bad_color = "1.0"  # white background
-colormap = gu.Colormap(bad_color=bad_color)
-my_cmap = colormap.cmap
+my_cmap = ColormapFactory(bad_color=bad_color).generate_cmap()
 
 #############
 # load data #
