+++ scripts/preprocessing/bcdi_read_data_P10.py
@@ -21,6 +21,8 @@ import matplotlib.pyplot as plt
 import os
 from scipy.io import savemat
 import sys
+
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 import bcdi.utils.utilities as util
 from bcdi.experiment.detector import create_detector
@@ -168,8 +170,7 @@ def main(parameters):
         bad_color = "0.7"
     else:
         bad_color = "1.0"  # white background
-    colormap = gu.Colormap(bad_color=bad_color)
-    my_cmap = colormap.cmap
+    my_cmap = ColormapFactory(bad_color=bad_color).generate_cmap()
 
     #######################
     # Initialize detector #
