+++ scripts/publication/bcdi_diffpattern_from_reconstruction.py
@@ -18,6 +18,8 @@ import pathlib
 from scipy.ndimage.measurements import center_of_mass
 import tkinter as tk
 from tkinter import filedialog
+
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 import bcdi.postprocessing.postprocessing_utils as pu
 import bcdi.utils.utilities as util
@@ -186,8 +188,7 @@ if grey_background:
 else:
     bad_color = "1.0"  # white background
 
-colormap = gu.Colormap(bad_color=bad_color)
-my_cmap = colormap.cmap
+my_cmap = ColormapFactory(bad_color=bad_color).generate_cmap()
 
 labels = ("Qx", "Qz", "Qy")
