+++ scripts/publication/bcdi_plot_strain.py
@@ -16,6 +16,7 @@ from scipy.ndimage.measurements import center_of_mass
 import tkinter as tk
 from tkinter import filedialog
 
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 import bcdi.utils.utilities as util
 import bcdi.utils.validation as valid
@@ -106,8 +107,7 @@ if grey_background:
     bad_color = "0.7"
 else:
     bad_color = "1.0"  # white background
-colormap = gu.Colormap(bad_color=bad_color)
-my_cmap = colormap.cmap
+my_cmap = ColormapFactory(bad_color=bad_color).generate_cmap()
 
 #########################
 # check some parameters #
