+++ scripts/xcca/bcdi_view_ccf_map.py
@@ -13,6 +13,8 @@ from matplotlib import pyplot as plt
 import tkinter as tk
 from tkinter import filedialog
 import sys
+
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 import bcdi.utils.utilities as util
 
@@ -151,8 +153,7 @@ def press_key(event):
 # define colormap #
 ###################
 bad_color = "1.0"  # white background
-colormap = gu.Colormap(bad_color=bad_color)
-my_cmap = colormap.cmap
+my_cmap = ColormapFactory(bad_color=bad_color).generate_cmap()
 plt.ion()
 
 ###################################
