+++ scripts/xcca/bcdi_xcca_3D_map_polar.py
@@ -18,6 +18,8 @@ from tkinter import filedialog
 import gc
 import multiprocessing as mp
 import sys
+
+from bcdi.graph.colormap import ColormapFactory
 import bcdi.graph.graph_utils as gu
 import bcdi.utils.utilities as util
 import bcdi.xcca.xcca_utils as xcca
@@ -125,8 +127,7 @@ def main(calc_self, user_comment):
     # define colormap #
     ###################
     bad_color = "1.0"  # white background
-    colormap = gu.Colormap(bad_color=bad_color)
-    my_cmap = colormap.cmap
+    my_cmap = ColormapFactory(bad_color=bad_color).generate_cmap()
     plt.ion()
 
     ###################################
