+++ tests/utils/test_utilities.py
@@ -272,37 +272,6 @@ class TestInRange(unittest.TestCase):
         self.assertTrue(util.in_range(point=(0, 0, 119), extent=self.extent))
 
 
-class TestIsFloat(unittest.TestCase):
-    """
-    Tests on the function utilities.is_float.
-
-    def is_float(string)
-    """
-
-    def test_string_float(self):
-        self.assertTrue(util.is_float("12.0"))
-
-    def test_string_int(self):
-        self.assertTrue(util.is_float("12"))
-
-    def test_string_complex(self):
-        self.assertFalse(util.is_float("12 + 1j"))
-
-    def test_string_none(self):
-        self.assertFalse(util.is_float("None"))
-
-    def test_string_not_numeric(self):
-        self.assertFalse(util.is_float("abc"))
-
-    def test_none(self):
-        with self.assertRaises(TypeError):
-            util.is_float(None)
-
-    def test_array(self):
-        with self.assertRaises(TypeError):
-            util.is_float(np.ones(3))
-
-
 class TestGaussianWindow(unittest.TestCase):
     """
     Tests on the function utilities.gaussian_window.
@@ -402,7 +371,6 @@ class TestNdarrayToList(unittest.TestCase):
 
 if __name__ == "__main__":
     run_tests(TestInRange)
-    run_tests(TestIsFloat)
     run_tests(TestFindFile)
     run_tests(TestGaussianWindow)
     run_tests(TestUnpackArray)
